languageserver: Add RunWithStdio function, move main.go to cmd/languageserver
